vrraiment finis

+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: nmontard <nmontard@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/10 10:02:06 by nmontard          #+#    #+#             */
-/*   Updated: 2025/11/16 11:00:07 by nmontard         ###   ########.fr       */
+/*   Updated: 2025/11/17 14:10:13 by nmontard         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,7 +33,7 @@ void	*ft_memcpy(void *dest, const void *src, size_t n);
 void	*ft_memset(void *s, int c, size_t n);
 char	*ft_strchr(char *str, char c);
 size_t	ft_strlcat(char *dest, const char *src, size_t size);
-int		ft_strlen(const char *str);
+size_t	ft_strlen(const char *str);
 int		ft_tolower(int c);
 int		ft_toupper(int c);
 void	*ft_memmove(void *dest, const void *src, size_t n);
